wip fetch cloud functions -- currently cors errors
