+++ package.json
@@ -21,7 +21,6 @@
     "firebase": "^7.14.2",
     "firebase-admin": "^8.11.0",
     "firebase-functions": "^3.6.1",
-    "gapi-client": "^0.0.3",
     "googleapis": "^49.0.0",
     "history": "^4.7.2",
     "http2": "^3.3.7",
@@ -30,6 +29,7 @@
     "mobx-react": "^6.2.2",
     "mobx-react-router": "^4.0.5",
     "moment": "2.24.0",
+    "node-fetch": "^2.6.0",
     "pubsub-js": "^1.7.0",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
@@ -51,7 +51,6 @@
     "deploy": "gh-pages -b master -d build"
   },
   "devDependencies": {
-    "@types/gapi": "^0.0.39",
     "@types/jest": "^25.2.1",
     "@types/node": "^13.13.4",
     "@types/react": "^16.7.13",
